build: extract `git-branchless-undo` crate
