+++ git-branchless-undo/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+edition = "2021"
+name = "git-branchless-undo"
+version = "0.7.0-pre.1"
+
+[dependencies]
+cursive = { version = "0.20.0", default-features = false, features = [
+  "crossterm-backend",
+] }
+eyre = "0.6.8"
+lib = { package = "git-branchless-lib", version = "0.7.0-pre.1", path = "../git-branchless-lib" }
+git-branchless-revset = { version = "0.7.0-pre.1", path = "../git-branchless-revset" }
+git-branchless-smartlog = { version = "0.7.0-pre.1", path = "../git-branchless-smartlog" }
+tracing = "0.1.37"
+cursive_buffered_backend = "0.6.1"
+git-record = { version = "0.3.0", path = "../git-record" }
+
+[dev-dependencies]
+insta = "1.21.0"
