+++ git-branchless/Cargo.toml
@@ -41,6 +41,7 @@ git-branchless-opts = { version = "0.7.0-pre.1", path = "../git-branchless-opts"
 git-branchless-revset = { version = "0.7.0-pre.1", path = "../git-branchless-revset" }
 git-branchless-smartlog = { version = "0.7.0-pre.1", path = "../git-branchless-smartlog" }
 git-branchless-test = { version = "0.7.0-pre.1", path = "../git-branchless-test" }
+git-branchless-undo = { version = "0.7.0-pre.1", path = "../git-branchless-undo" }
 git-record = { version = "0.3", path = "../git-record" }
 glob = "0.3.0"
 itertools = "0.10.3"
