+++ git-branchless/src/lib.rs
@@ -21,4 +21,3 @@
 #![allow(clippy::too_many_arguments, clippy::blocks_in_if_conditions)]
 
 pub mod commands;
-pub mod tui;
